+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## v2.0.1
+
+### `@liveblocks/node`
+
+- Fix type signatures of `client.identifyUser()` and `client.prepareSession()`
+  to require `userInfo` if it's mandatory according to your global `UserMeta`
+  type definition.
+
 ## v2.0.0
 
 This major release marks the maturity of Liveblocks. It contains new products